Move logger function out of the way
